refactor: migrate MFRouter on typescript
