+++ demo/3000-home/pages/_app.real.js
@@ -22,7 +22,7 @@ function MyApp({ Component, pageProps }) {
   useEffect(() => {
     if (
       // pathname !== '/[...federatedPage]' &&
-      !window.mf_nextjs.isFederatedPathname(pathname) &&
+      !window.mf_router.isFederatedPathname(pathname) &&
       MenuComponent !== AppMenu
     ) {
       setMenuComponent(() => AppMenu);
