+++ demo/3002-checkout/src/pages/_app.real.js
@@ -30,7 +30,7 @@ function MyApp({ Component, pageProps }) {
   useEffect(() => {
     if (
       // pathname !== '/[...federatedPage]' &&
-      !window.mf_nextjs.isFederatedPathname(pathname) &&
+      !window.mf_router.isFederatedPathname(pathname) &&
       MenuComponent !== AppMenu
     ) {
       setMenuComponent(() => AppMenu);
