+++ demo/package.json
@@ -9,9 +9,9 @@
   ],
   "scripts": {
     "clear": "cd 3000-home && rm -rf node_modules && rm -rf .next && cd ../3001-shop && rm -rf node_modules && rm -rf .next && cd ../3002-checkout && rm -rf node_modules && rm -rf .next",
-    "dev": "concurrently \"yarn sync-pkg -w\" \"cd 3000-home; npm run dev\" \"cd 3001-shop; npm run dev\" \"cd 3002-checkout; npm run dev\"",
-    "build": "yarn sync-pkg && concurrently \"cd 3000-home; npm run build\" \"cd 3001-shop; npm run build\" \"cd 3002-checkout; npm run build\"",
-    "start": "yarn sync-pkg && concurrently \"cd 3000-home; npm run start\" \"cd 3001-shop; npm run start\" \"cd 3002-checkout; npm run start\"",
+    "dev": "concurrently -n _,home,shop,checkout \"yarn sync-pkg -w\" \"cd 3000-home; npm run dev\" \"cd 3001-shop; npm run dev\" \"cd 3002-checkout; npm run dev\"",
+    "build": "yarn sync-pkg && concurrently -n home,shop,checkout \"cd 3000-home; npm run build\" \"cd 3001-shop; npm run build\" \"cd 3002-checkout; npm run build\"",
+    "start": "yarn sync-pkg && concurrently -n home,shop,checkout \"cd 3000-home; npm run start\" \"cd 3001-shop; npm run start\" \"cd 3002-checkout; npm run start\"",
     "sync-pkg": "cpx \"../lib/**\" ./tmp/nextjs-mf/lib"
   },
   "dependencies": {
